requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.3
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-multipart==0.0.9